backend/requirements.txt
beautifulsoup4==4.14.3
celery==5.6.3
django==5.2.12
fastapi==0.135.2
google-cloud-vision==3.13.0
opencv-python==4.13.0.92
openpyxl==3.1.5
pandas==2.3.3
pillow==12.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pydantic_settings==2.13.1
pytesseract==0.3.13
requests==2.33.0
redis==7.4.0
scrapy==2.14.2
selenium==4.41.0
sqlalchemy==2.0.48
Telethon==1.42.0
Tesseract==0.1.3
torch==2.11.0
torchaudio==2.11.0
torchvision==0.26.0
uvicorn==0.42.0
wheel==0.46.3